+++ src/views/components/UI/organisms/Header.tsx
@@ -11,7 +11,7 @@ import {
   MenuIcon,
   LogoutBoxIcon,
   CloseIcon,
-  RichCheckbox,
+  Checkbox,
   Label,
 } from '@components';
 import { UtilisateurReadModel } from '@modules/utilisateur/récupérer/UtilisateurReadModel';
@@ -189,7 +189,7 @@ type MainMenuProps = {
 };
 const MainMenu = ({ children }: MainMenuProps) => (
   <>
-    <RichCheckbox id="menu-toggle" className="hidden" />
+    <Checkbox id="menu-toggle" className="hidden" />
     <Label className="absolute top-3 right-2 text-xl lg:hidden" htmlFor="menu-toggle">
       <MenuIcon className="menu-open" />
       <CloseIcon className="menu-close hidden" />
