+++ src/views/components/UI/organisms/ProjectList.tsx
@@ -17,7 +17,7 @@ import {
   LinkButton,
   Tile,
   PaginationPanel,
-  RichCheckbox,
+  Checkbox,
 } from '@components';
 import { afficherDate } from '@views/helpers';
 
@@ -137,7 +137,7 @@ export const ProjectList = ({
       <div className="p-5 flex items-center">
         {displaySelection && (
           <>
-            <RichCheckbox
+            <Checkbox
               id="allProjects"
               onChange={(e) => toggleSelectAllPage(e.target.checked)}
               checked={selectedIds.length === projects.items.length}
@@ -156,7 +156,7 @@ export const ProjectList = ({
               <div className="flex flex-col gap-2 mb-4">
                 <div className="flex flex-col md:flex-row gap-2">
                   {displaySelection && (
-                    <RichCheckbox
+                    <Checkbox
                       id={project.id}
                       onChange={(e) => toggleSelected(project.id, e.target.checked)}
                       value={project.id}
